code refactoring, added support for showing active track in tracklist
