+++ src/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from './use-data-action';
 export * from './use-data-load';
 export * from './use-entity-missing';
-export * from './use-media-playback';
+export * from './use-media-collection-playback';
+export * from './use-media-track-playback';
